Seems to be working. Not fully tested

+++ empac/joystick.c
@@ -133,7 +133,7 @@ void joyCalculateOutput(JoystickData2D *jd)
 
 			//Calculate the axis output
 			jd->axis[thisAxis].output 
-			=	(int16_t)((offCentre*jd->axis[thisAxis].outputMax)
+			=	((offCentre*jd->axis[thisAxis].outputMax)
 			/	(jd->axis[thisAxis].rawMax
 			-	jd->axis[thisAxis].rawCnt
 			-	jd->axis[thisAxis].deadzone));
